requirements.txt
openai
selenium
psycopg2-binary
streamlit
mailjet_rest
requests
python-dotenv
webdriver-manager
